Organize webdriver commands
